+++ core/src/onsenui.d.ts
@@ -270,7 +270,7 @@ declare namespace ons {
      * @return {Boolean}
      */
     function isIPhoneX(): boolean;
-    
+
     /**
      * @description Returns whether the device is iPad
      * @return {Boolean}
@@ -742,7 +742,12 @@ declare namespace ons {
      * @description Default options object. Attributes have priority over this property.
      */
     options: NavigatorOptions;
-
+    /**
+     * @param {Number} index Decimal ratio of the current swipe.
+     * @param {Object} animationOptions Object containing duration and timing.
+     * @description Hook called whenever the user slides the navigator.
+     **/
+    onSwipe?: Function;
   }
 
   /**
@@ -859,6 +864,12 @@ declare namespace ons {
      * @return Resolves to the new page element
      */
     load(page: any, options?: SplitterSideOptions): Promise<HTMLElement>;
+    /**
+     * @param {Number} index Decimal ratio of the current swipe.
+     * @param {Object} animationOptions Object containing duration and timing.
+     * @description Hook called whenever the user slides the splitter.
+     **/
+    onSwipe?: Function;
   }
 
   interface OnsLazyRepeatElement extends HTMLElement {