+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGELOG
 dev
 ----
 
+### New Features
+
+ * ons-navigator: Added `onSwipe` property function that gets the swiped distance on drag.
+
 ### Bug Fixes
 
  * ons-splitter: Fixed small glitch in 'reveal' animation.