+++ CHANGELOG.md
@@ -16,6 +16,7 @@ dev
  * ons-select: Fixed custom modifier class from `select-*` to `select--*`.
  * ons-tabbar: Fixed [#2253](https://github.com/OnsenUI/OnsenUI/issues/2253).
  * ons-tabbar: Position 'auto' and border visibility now cares about autostyling instead of the actual platform.
+ * angular1: Remove `element.scope()` call that depends on AngularJS Debug Data.
 
 2.8.1
 ----